requirement.txt
faiss-cpu==1.7.2
glob2==0.7
tqdm==4.64.1
numpy==1.22.0
opencv-python==4.6.0.66
imageio==2.22.0
matplotlib==3.5.3
Pillow==9.3.0
PyYAML==6.0
requests==2.28.1
scipy==1.7.3
scikit-image==0.19.3
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@6ae6ff9cf73221ce60617ef4658b4892b986ba9d
chumpy
pytorch-lightning==1.9.4
einops
torchgeometry
tensorboard==2.12.0